+++ runner/src/index.ts
@@ -55,4 +55,5 @@ export {
   saveToBlobby,
   setBlobbyServerUrl,
 } from "./blobby-storage.ts";
+export { tsToExports } from "./local-build.ts";
 export { addCommonIDfromObjectID } from "./utils.ts";
